+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/TableConditionPanel.java
@@ -150,7 +150,7 @@ public abstract class TableConditionPanel<C> extends JPanel {
 			panelItems.get(0).value().requestInputFocus();
 		}
 		else if (!panelItems.isEmpty()) {
-			Dialogs.selectionDialog(panelItems)
+			Dialogs.listSelectionDialog(panelItems)
 							.owner(dialogOwner)
 							.title(MESSAGES.getString("select_condition"))
 							.selectSingle()