+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/DefaultTextFieldBuilder.java
@@ -20,7 +20,7 @@ package is.codion.swing.common.ui.component.text;
 
 import is.codion.common.value.Value;
 import is.codion.swing.common.ui.component.value.ComponentValue;
-import is.codion.swing.common.ui.dialog.SelectionDialogBuilder.SingleSelector;
+import is.codion.swing.common.ui.dialog.ListSelectionDialogBuilder.SingleSelector;
 import is.codion.swing.common.ui.key.KeyEvents;
 
 import javax.swing.AbstractAction;