+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/SelectionDialogBuilder.java
@@ -18,80 +18,31 @@
  */
 package is.codion.swing.common.ui.dialog;
 
-import javax.swing.JComponent;
-import java.awt.Dimension;
 import java.util.Collection;
-import java.util.Optional;
 
 /**
  * A builder for a selection dialog.
  * @param <T> the value type
  */
-public interface SelectionDialogBuilder<T> extends DialogBuilder<SelectionDialogBuilder<T>> {
+public interface SelectionDialogBuilder<T, B extends SelectionDialogBuilder<T, B>> extends DialogBuilder<B> {
 
 	/**
 	 * @param defaultSelection the item selected by default
 	 * @return this SelectionDialogBuilder instance
 	 * @throws IllegalArgumentException in case the selection values do not contain the default selection item
 	 */
-	SelectionDialogBuilder<T> defaultSelection(T defaultSelection);
+	B defaultSelection(T defaultSelection);
 
 	/**
 	 * @param defaultSelection the items selected by default
 	 * @return this SelectionDialogBuilder instance
 	 * @throws IllegalArgumentException in case the selection values do not contain the default selection items
 	 */
-	SelectionDialogBuilder<T> defaultSelection(Collection<T> defaultSelection);
+	B defaultSelection(Collection<T> defaultSelection);
 
 	/**
 	 * @param allowEmptySelection if true then the dialog accepts an empty selection, default false
 	 * @return this SelectionDialogBuilder instance
 	 */
-	SelectionDialogBuilder<T> allowEmptySelection(boolean allowEmptySelection);
-
-	/**
-	 * @param dialogSize the preferred dialog size
-	 * @return this SelectionDialogBuilder instance
-	 */
-	SelectionDialogBuilder<T> dialogSize(Dimension dialogSize);
-
-	/**
-	 * @return the selected value, {@link Optional#empty()} if none was selected
-	 * @throws is.codion.common.model.CancelException in case the user cancelled
-	 */
-	Optional<T> selectSingle();
-
-	/**
-	 * @return the selected values, an empty Collection if none was selected
-	 * @throws is.codion.common.model.CancelException in case the user cancelled
-	 */
-	Collection<T> select();
-
-	/**
-	 * Provides the user with the ability to select a single value.
-	 * @param <T> the value type
-	 */
-	interface SingleSelector<T> {
-
-		/**
-		 * @param dialogOwner the dialog owner
-		 * @return the selected value, an empty Optional if nothing was selected
-		 * @throws is.codion.common.model.CancelException in case the user cancelled
-		 */
-		Optional<T> select(JComponent dialogOwner);
-	}
-
-	/**
-	 * Provides the user with the ability to select one or more values.
-	 * @param <T> the value type
-	 */
-	interface MultiSelector<T> {
-
-		/**
-		 * @param dialogOwner the dialog owner
-		 * @return the selected values, an empty Collections if nothing was selected
-		 * @throws is.codion.common.model.CancelException in case the user cancelled
-		 */
-		Collection<T> select(JComponent dialogOwner);
-	}
+	B allowEmptySelection(boolean allowEmptySelection);
 }