+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTablePanel.java
@@ -638,7 +638,7 @@ public class EntityTablePanel extends JPanel {
 						.map(attribute -> tableModel.entityDefinition().attributes().definition(attribute))
 						.sorted(new AttributeDefinitionComparator())
 						.collect(toList());
-		Dialogs.selectionDialog(sortedDefinitions)
+		Dialogs.listSelectionDialog(sortedDefinitions)
 						.owner(this)
 						.selectSingle()
 						.map(AttributeDefinition::attribute)